Fix not re-initializing event handlers
We had an issue, where we registered event handlers only for the first
initialization. Sometimes consumers have to re-initialize the RyePay.
This happens, for example, on a single page applications, where payment
form is rendered on a separate step, and user navigates away from the
payment form after they see it, and then returns to payment form.
In this case all event handlers are initialized only for the first
time user sees the form. On the second init event handlers are not
registered, and stale callbacks are executed. One particular issue
with stale callbacks was that we tried to update React state after
changes in forms, and the `set` functions were stale (captured `set`
functions were from the previous instance of the component).